refactor: reorganize pool physics events into separate modules
